Ignore import/no-unresolved, add clarification comment.
The CI would complain about the missing config if the rule was not ignored. It's because at build-time, the config comes from an environment variable so it's not committed directly to the repository. In the case of local development, though, developers should copy the example config file and adjust the option to their needs.